Refactor: loggingdbをloggingdriverに変更
