+++ internal/infrastructure/rdb/driver/loggingdriver/provider_test.go
@@ -1,4 +1,4 @@
-package loggingdb
+package loggingdriver
 
 import (
 	"context"
